+++ src/main/java/net/fneifnox/custommobattributes/mixin/movement/SquidSpeedMixin.java
@@ -10,15 +10,9 @@ import static net.fneifnox.custommobattributes.CustomMobAttributes.CONFIG;
 @Mixin(SquidEntity.class)
 public class SquidSpeedMixin {
 
-    @Redirect(
-            method = "tickMovement()V",
-            at = @At(
-                    value = "INVOKE",
-                    target = "Lnet/minecraft/entity/passive/SquidEntity;setVelocity(DDD)V"
-            )
-    )
+    @Redirect(method = "tickMovement()V", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/passive/SquidEntity;setVelocity(DDD)V"))
     private void redirectSetVelocity(SquidEntity instance, double x, double y, double z) {
-        float speedMultiplier = CONFIG.speedMultiplierForSquid() * CONFIG.speedMultiplierForAll();
+        double speedMultiplier = CONFIG.speedMultiplierForSquid() * CONFIG.speedMultiplierForAll();
         instance.setVelocity(x * speedMultiplier, y * speedMultiplier, z * speedMultiplier);
     }
 }
